requirements.txt
requests>=2.28.2
python-dotenv>=0.21.1
Flask-Caching>=2.0.2
Flask>=2.2.3
Flask-Cors>=3.0.10
redis>=4.5.4
